master backend commucation

+++ master/conhash.cc
@@ -1,11 +1,10 @@
 #include "conhash.h"
 
-unsigned hash_str(const char* s)
-{
-   unsigned h = FIRSTH;
-   while (*s) {
-     h = (h * A) ^ (s[0] * B);
-     s++;
-   }
-   return h%MAX_KEY;
+int hash_str(const char *s) {
+  unsigned h = FIRSTH;
+  while (*s) {
+    h = (h * A) ^ (s[0] * B);
+    s++;
+  }
+  return (int)(h % MAX_KEY);
 }
